Updated to Ep. 13
* Added Layers, Groups, Timer and more
* Created Optimized DLL for better performance
* Some JNI code optimizations

+++ native/Sparky4j-core/src/maths/maths_funcs.h
@@ -7,7 +7,7 @@ namespace sparky {
 	namespace maths {
 
 		inline float toRadians(float degrees) {
-			return degrees * (M_PI / 180.0f);
+			return (float) (degrees * (M_PI / 180.0f));
 		}
 
 	}
